Update main.lua; remove tile creation bug
With prior code, many "Tile" objects would be created and added to the "trackTracker" table when holding down the mouse, and "Tile" objects could be created in grid squares already displaying tracks. The "track_tracker" table of tables corresponds to the columns and rows, and it checks whether a "Tile" object exists for the corresponding position. Bug fixed, and overall memory and efficiency is improved.